+++ modules/access/screen/screen.c
@@ -301,7 +301,7 @@ static int Control( demux_t *p_demux, int i_query, va_list args )
 
         case DEMUX_GET_PTS_DELAY:
             pi64 = (int64_t*)va_arg( args, int64_t * );
-            *pi64 = (int64_t)var_GetInteger( p_demux, "screen-caching" ) *1000;
+            *pi64 = var_GetInteger( p_demux, "screen-caching" ) *1000;
             return VLC_SUCCESS;
 
         case DEMUX_GET_TIME:
